[REFACTORING] Relocate Vacation models in vacation package
